Add --generate-api-docs to build script
Add a new option `--generate-api-docs` that changes the build script to
only do what is required to build the API documentation, skipping the
process of building the full Atom binaries.

Also fixes the script so argument handling is done before the bootstrap
process, saving quite a bit of time if only called with `--help` or with
invalid arguments.